+++ src/main/java/com/factura/FacturaElectronica/Bo/KardexBo.java
@@ -1,7 +1,5 @@
 package com.factura.FacturaElectronica.Bo;
 
-import java.math.BigDecimal;
-
 import com.factura.FacturaElectronica.modelo.Inventario;
 import com.factura.FacturaElectronica.modelo.Kardex;
 import com.factura.FacturaElectronica.modelo.Usuario;
@@ -10,7 +8,7 @@ public interface KardexBo {
 
 	void agregar(Kardex kardex);
 
-	void entrada(Inventario inventario, BigDecimal cantidad, String observacion, String consecutivo, String tipo, String motivo,Usuario usuario);
-	void salida(Inventario inventario, BigDecimal cantidad, String observacion, String consecutivo, String tipo, String motivo,Usuario usuario);
+	void entrada(Inventario inventario, Double cantidad, String observacion, String consecutivo, String tipo, String motivo,Usuario usuario);
+	void salida(Inventario inventario, Double cantidad, String observacion, String consecutivo, String tipo, String motivo,Usuario usuario);
 
 }
