+++ src/main/java/com/factura/FacturaElectronica/Dao/ArticuloDao.java
@@ -1,7 +1,5 @@
 package com.factura.FacturaElectronica.Dao;
 
-import java.math.BigDecimal;
-
 import com.factura.FacturaElectronica.modelo.Articulo;
 import com.factura.FacturaElectronica.modelo.Empresa;
 
@@ -19,8 +17,8 @@ public interface ArticuloDao {
 
 	Articulo buscarPorCodigoYEmpresa(String codigo, Empresa empresa);
 	
-	BigDecimal porcentanjeDeGanancia(BigDecimal costo, BigDecimal iva, BigDecimal precio);
+	Double porcentanjeDeGanancia(Double costo, Double iva, Double precio);
 	
-	BigDecimal costoPromedio(BigDecimal costoActual , BigDecimal costoNuevo,BigDecimal cantidadActual,BigDecimal cantidadNueva);
+	Double costoPromedio(Double costoActual , Double costoNuevo,Double cantidadActual,Double cantidadNueva);
 
 }
