+++ src/main/java/com/factura/FacturaElectronica/Dao/InventarioDao.java
@@ -1,7 +1,5 @@
 package com.factura.FacturaElectronica.Dao;
 
-import java.math.BigDecimal;
-
 import com.factura.FacturaElectronica.modelo.Articulo;
 import com.factura.FacturaElectronica.modelo.Inventario;
 
@@ -17,7 +15,7 @@ public interface InventarioDao {
 
 	Inventario buscarPorArticulo(Articulo articulo);
 	Inventario findByArticuloAndEstado(Articulo articulo,String estado);
-	BigDecimal getTotalCosto(Inventario inventario, BigDecimal cantidad);
+	Double getTotalCosto(Inventario inventario, Double cantidad);
