+++ src/main/java/com/factura/FacturaElectronica/modelo/Inventario.java
@@ -1,7 +1,6 @@
 package com.factura.FacturaElectronica.modelo;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
 import java.util.Date;
 import java.util.HashSet;
 import java.util.Set;
@@ -42,13 +41,13 @@ public class Inventario implements Serializable {
 	private Integer						id;
 
 	@Column(name = "cantidad")
-	private BigDecimal							cantidad;
+	private Double						cantidad;
 
 	@Column(name = "minimo")
-	private Double							minimo;
+	private Double						minimo;
 
 	@Column(name = "maximo")
-	private Double							maximo;
+	private Double						maximo;
 
 	@Column(name = "estado")
 	private String						estado;
@@ -85,8 +84,7 @@ public class Inventario implements Serializable {
 
 	}
 
-
-	public Inventario(Integer id, BigDecimal cantidad, Double minimo, Double maximo, String estado, Date created_at, Date updated_at, Articulo articulo, Usuario usuario, Set<Kardex> kardexs) {
+	public Inventario(Integer id, Double cantidad, Double minimo, Double maximo, String estado, Date created_at, Date updated_at, Articulo articulo, Usuario usuario, Set<Kardex> kardexs) {
 		super();
 		this.id = id;
 		this.cantidad = cantidad;
@@ -100,8 +98,6 @@ public class Inventario implements Serializable {
 		this.kardexs = kardexs;
 	}
 
-
-
 	public Integer getId() {
 		return id;
 	}
@@ -110,21 +106,13 @@ public class Inventario implements Serializable {
 		this.id = id;
 	}
 
-	
-
-	
-	public BigDecimal getCantidad() {
+	public Double getCantidad() {
 		return cantidad;
 	}
 
-
-
-	
-	public void setCantidad(BigDecimal cantidad) {
+	public void setCantidad(Double cantidad) {
 		this.cantidad = cantidad;
 	}
-
-
 
 	public Double getMinimo() {
 		return minimo;
