+++ src/main/java/com/factura/FacturaElectronica/web/command/FacturaEsperaCommand.java
@@ -1,6 +1,5 @@
 package com.factura.FacturaElectronica.web.command;
 
-import java.math.BigDecimal;
 import java.util.Date;
 
 import com.factura.FacturaElectronica.modelo.Cliente;
@@ -16,7 +15,6 @@ public class FacturaEsperaCommand {
 	private Date			fechaCredito;
 
 	private String		numeroConsecutivo;
-	                  
 
 	private Date			fechaEmision;
 
@@ -36,45 +34,45 @@ public class FacturaEsperaCommand {
 
 	private String		comanda;
 
-	private BigDecimal		subTotal;
+	private Double		subTotal;
 
-	private BigDecimal		totalTransporte;
+	private Double		totalTransporte;
 
-	private BigDecimal		total;
+	private Double		total;
 
-	private BigDecimal		totalServGravados;
+	private Double		totalServGravados;
 
-	private BigDecimal		totalServExentos;
+	private Double		totalServExentos;
 
-	private BigDecimal		totalMercanciasGravadas;
+	private Double		totalMercanciasGravadas;
 
-	private BigDecimal		totalMercanciasExentas;
+	private Double		totalMercanciasExentas;
 
-	private BigDecimal		totalGravado;
-	private BigDecimal		totalExento;
+	private Double		totalGravado;
+	private Double		totalExento;
 
-	private BigDecimal		totalVenta;
+	private Double		totalVenta;
 
-	private BigDecimal		totalDescuentos;
-	private BigDecimal		totalVentaNeta;
+	private Double		totalDescuentos;
+	private Double		totalVentaNeta;
 
-	private BigDecimal		totalImpuesto;
+	private Double		totalImpuesto;
 
-	private BigDecimal		totalComprobante;
+	private Double		totalComprobante;
 
-	private BigDecimal		totalEfectivo;
+	private Double		totalEfectivo;
 
-	private BigDecimal		totalTarjeta;
+	private Double		totalTarjeta;
 
-	private BigDecimal		totalBanco;
+	private Double		totalBanco;
 
-	private BigDecimal		totalCredito;
+	private Double		totalCredito;
 
-	private BigDecimal		montoCambio;
+	private Double		montoCambio;
 
-	private BigDecimal		totalCambio;
+	private Double		totalCambio;
 
-	private BigDecimal		totalCambioPagar;
+	private Double		totalCambioPagar;
 
 	private String		codigoMoneda;
 
@@ -137,415 +135,332 @@ public class FacturaEsperaCommand {
 		this.usuarioCreacion = factura.getUsuarioCreacion();
 	}
 
-	
 	public Integer getId() {
 		return id;
 	}
 
-	
 	public void setId(Integer id) {
 		this.id = id;
 	}
 
-	
 	public Date getFechaCredito() {
 		return fechaCredito;
 	}
 
-	
 	public void setFechaCredito(Date fechaCredito) {
 		this.fechaCredito = fechaCredito;
 	}
 
-	
 	public String getNumeroConsecutivo() {
 		return numeroConsecutivo;
 	}
 
-	
 	public void setNumeroConsecutivo(String numeroConsecutivo) {
 		this.numeroConsecutivo = numeroConsecutivo;
 	}
 
-	
 	public Date getFechaEmision() {
 		return fechaEmision;
 	}
 
-	
 	public void setFechaEmision(Date fechaEmision) {
 		this.fechaEmision = fechaEmision;
 	}
 
-	
 	public String getCondicionVenta() {
 		return condicionVenta;
 	}
 
-	
 	public void setCondicionVenta(String condicionVenta) {
 		this.condicionVenta = condicionVenta;
 	}
 
-	
 	public Integer getPlazoCredito() {
 		return plazoCredito;
 	}
 
-	
 	public void setPlazoCredito(Integer plazoCredito) {
 		this.plazoCredito = plazoCredito;
 	}
 
-	
 	public String getTipoDoc() {
 		return tipoDoc;
 	}
 
-	
 	public void setTipoDoc(String tipoDoc) {
 		this.tipoDoc = tipoDoc;
 	}
 
-	
 	public String getMedioPago() {
 		return medioPago;
 	}
 
-	
 	public void setMedioPago(String medioPago) {
 		this.medioPago = medioPago;
 	}
 
-	
 	public String getNombreFactura() {
 		return nombreFactura;
 	}
 
-	
 	public void setNombreFactura(String nombreFactura) {
 		this.nombreFactura = nombreFactura;
 	}
 
-	
 	public String getDireccion() {
 		return direccion;
 	}
 
-	
 	public void setDireccion(String direccion) {
 		this.direccion = direccion;
 	}
 
-	
 	public String getNota() {
 		return nota;
 	}
 
-	
 	public void setNota(String nota) {
 		this.nota = nota;
 	}
 
-	
 	public String getComanda() {
 		return comanda;
 	}
 
-	
 	public void setComanda(String comanda) {
 		this.comanda = comanda;
 	}
 
-	
-	public BigDecimal getSubTotal() {
+	public Double getSubTotal() {
 		return subTotal;
 	}
 
-	
-	public void setSubTotal(BigDecimal subTotal) {
+	public void setSubTotal(Double subTotal) {
 		this.subTotal = subTotal;
 	}
 
-	
-	public BigDecimal getTotalTransporte() {
+	public Double getTotalTransporte() {
 		return totalTransporte;
 	}
 
-	
-	public void setTotalTransporte(BigDecimal totalTransporte) {
+	public void setTotalTransporte(Double totalTransporte) {
 		this.totalTransporte = totalTransporte;
 	}
 
-	
-	public BigDecimal getTotal() {
+	public Double getTotal() {
 		return total;
 	}
 
-	
-	public void setTotal(BigDecimal total) {
+	public void setTotal(Double total) {
 		this.total = total;
 	}
 
-	
-	public BigDecimal getTotalServGravados() {
+	public Double getTotalServGravados() {
 		return totalServGravados;
 	}
 
-	
-	public void setTotalServGravados(BigDecimal totalServGravados) {
+	public void setTotalServGravados(Double totalServGravados) {
 		this.totalServGravados = totalServGravados;
 	}
 
-	
-	public BigDecimal getTotalServExentos() {
+	public Double getTotalServExentos() {
 		return totalServExentos;
 	}
 
-	
-	public void setTotalServExentos(BigDecimal totalServExentos) {
+	public void setTotalServExentos(Double totalServExentos) {
 		this.totalServExentos = totalServExentos;
 	}
 
-	
-	public BigDecimal getTotalMercanciasGravadas() {
+	public Double getTotalMercanciasGravadas() {
 		return totalMercanciasGravadas;
 	}
 
-	
-	public void setTotalMercanciasGravadas(BigDecimal totalMercanciasGravadas) {
+	public void setTotalMercanciasGravadas(Double totalMercanciasGravadas) {
 		this.totalMercanciasGravadas = totalMercanciasGravadas;
 	}
 
-	
-	public BigDecimal getTotalMercanciasExentas() {
+	public Double getTotalMercanciasExentas() {
 		return totalMercanciasExentas;
 	}
 
-	
-	public void setTotalMercanciasExentas(BigDecimal totalMercanciasExentas) {
+	public void setTotalMercanciasExentas(Double totalMercanciasExentas) {
 		this.totalMercanciasExentas = totalMercanciasExentas;
 	}
 
-	
-	public BigDecimal getTotalGravado() {
+	public Double getTotalGravado() {
 		return totalGravado;
 	}
 
-	
-	public void setTotalGravado(BigDecimal totalGravado) {
+	public void setTotalGravado(Double totalGravado) {
 		this.totalGravado = totalGravado;
 	}
 
-	
-	public BigDecimal getTotalExento() {
+	public Double getTotalExento() {
 		return totalExento;
 	}
 
-	
-	public void setTotalExento(BigDecimal totalExento) {
+	public void setTotalExento(Double totalExento) {
 		this.totalExento = totalExento;
 	}
 
-	
-	public BigDecimal getTotalVenta() {
+	public Double getTotalVenta() {
 		return totalVenta;
 	}
 
-	
-	public void setTotalVenta(BigDecimal totalVenta) {
+	public void setTotalVenta(Double totalVenta) {
 		this.totalVenta = totalVenta;
 	}
 
-	
-	public BigDecimal getTotalDescuentos() {
+	public Double getTotalDescuentos() {
 		return totalDescuentos;
 	}
 
-	
-	public void setTotalDescuentos(BigDecimal totalDescuentos) {
+	public void setTotalDescuentos(Double totalDescuentos) {
 		this.totalDescuentos = totalDescuentos;
 	}
 
-	
-	public BigDecimal getTotalVentaNeta() {
+	public Double getTotalVentaNeta() {
 		return totalVentaNeta;
 	}
 
-	
-	public void setTotalVentaNeta(BigDecimal totalVentaNeta) {
+	public void setTotalVentaNeta(Double totalVentaNeta) {
 		this.totalVentaNeta = totalVentaNeta;
 	}
 
-	
-	public BigDecimal getTotalImpuesto() {
+	public Double getTotalImpuesto() {
 		return totalImpuesto;
 	}
 
-	
-	public void setTotalImpuesto(BigDecimal totalImpuesto) {
+	public void setTotalImpuesto(Double totalImpuesto) {
 		this.totalImpuesto = totalImpuesto;
 	}
 
-	
-	public BigDecimal getTotalComprobante() {
+	public Double getTotalComprobante() {
 		return totalComprobante;
 	}
 
-	
-	public void setTotalComprobante(BigDecimal totalComprobante) {
+	public void setTotalComprobante(Double totalComprobante) {
 		this.totalComprobante = totalComprobante;
 	}
 
-	
-	public BigDecimal getTotalEfectivo() {
+	public Double getTotalEfectivo() {
 		return totalEfectivo;
 	}
 
-	
-	public void setTotalEfectivo(BigDecimal totalEfectivo) {
+	public void setTotalEfectivo(Double totalEfectivo) {
 		this.totalEfectivo = totalEfectivo;
 	}
 
-	
-	public BigDecimal getTotalTarjeta() {
+	public Double getTotalTarjeta() {
 		return totalTarjeta;
 	}
 
-	
-	public void setTotalTarjeta(BigDecimal totalTarjeta) {
+	public void setTotalTarjeta(Double totalTarjeta) {
 		this.totalTarjeta = totalTarjeta;
 	}
 
-	
-	public BigDecimal getTotalBanco() {
+	public Double getTotalBanco() {
 		return totalBanco;
 	}
 
-	
-	public void setTotalBanco(BigDecimal totalBanco) {
+	public void setTotalBanco(Double totalBanco) {
 		this.totalBanco = totalBanco;
 	}
 
-	
-	public BigDecimal getTotalCredito() {
+	public Double getTotalCredito() {
 		return totalCredito;
 	}
 
-	
-	public void setTotalCredito(BigDecimal totalCredito) {
+	public void setTotalCredito(Double totalCredito) {
 		this.totalCredito = totalCredito;
 	}
 
-	
-	public BigDecimal getMontoCambio() {
+	public Double getMontoCambio() {
 		return montoCambio;
 	}
 
-	
-	public void setMontoCambio(BigDecimal montoCambio) {
+	public void setMontoCambio(Double montoCambio) {
 		this.montoCambio = montoCambio;
 	}
 
-	
-	public BigDecimal getTotalCambio() {
+	public Double getTotalCambio() {
 		return totalCambio;
 	}
 
-	
-	public void setTotalCambio(BigDecimal totalCambio) {
+	public void setTotalCambio(Double totalCambio) {
 		this.totalCambio = totalCambio;
 	}
 
-	
-	public BigDecimal getTotalCambioPagar() {
+	public Double getTotalCambioPagar() {
 		return totalCambioPagar;
 	}
 
-	
-	public void setTotalCambioPagar(BigDecimal totalCambioPagar) {
+	public void setTotalCambioPagar(Double totalCambioPagar) {
 		this.totalCambioPagar = totalCambioPagar;
 	}
 
-	
 	public String getCodigoMoneda() {
 		return codigoMoneda;
 	}
 
-	
 	public void setCodigoMoneda(String codigoMoneda) {
 		this.codigoMoneda = codigoMoneda;
 	}
 
-	
 	public Integer getEstado() {
 		return estado;
 	}
 
-	
 	public void setEstado(Integer estado) {
 		this.estado = estado;
 	}
 
-	
 	public Date getCreated_at() {
 		return created_at;
 	}
 
-	
 	public void setCreated_at(Date created_at) {
 		this.created_at = created_at;
 	}
 
-	
 	public Date getUpdated_at() {
 		return updated_at;
 	}
 
-	
 	public void setUpdated_at(Date updated_at) {
 		this.updated_at = updated_at;
 	}
 
-	
 	public Cliente getCliente() {
 		return cliente;
 	}
 
-	
 	public void setCliente(Cliente cliente) {
 		this.cliente = cliente;
 	}
 
-	
 	public Empresa getEmpresa() {
 		return empresa;
 	}
 
-	
 	public void setEmpresa(Empresa empresa) {
 		this.empresa = empresa;
 	}
 
-	
 	public Vendedor getVendedor() {
 		return vendedor;
 	}
 
-	
 	public void setVendedor(Vendedor vendedor) {
 		this.vendedor = vendedor;
 	}
 
-	
 	public Usuario getUsuarioCreacion() {
 		return usuarioCreacion;
 	}
 
-	
 	public void setUsuarioCreacion(Usuario usuarioCreacion) {
 		this.usuarioCreacion = usuarioCreacion;
 	}
 
-	
 }
